# apps/api/docs/swagger/front/openapi.bundle.yaml
openapi: 3.0.0
info:
  title: Nexus 前台 (Front) API
  version: 1.0.0
  description: |
    前台 API 文件。

    **除了 `/auth/*` 與 `/ping` 之外，所有端點都需要前台的 Bearer Token**
    （由 `POST /front/auth/login` 取得）。

    ⚠️ **後台（`/api/admin/*`）的 token 在這裡一律無效**——兩側各自使用不同的
    簽發 secret，簽章驗證就會失敗，不是權限不足。WebSocket 連線同理。
servers:
  - url: 'http://localhost:3000/api/front'
    description: 本地開發環境（前台 API）
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        使用**前台**的 Access Token 進行身份驗證（`POST /front/auth/login` 取得）

        格式: Bearer {accessToken}

        後台（`/api/admin/*`）簽出的 token 在這裡一律無效：
        兩側各自使用不同的 secret，簽章驗證就會失敗。
  schemas:
...
    ErrorResponse:
      type: object
      description: 通用錯誤回應格式
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: 操作是否成功
          example: false
        message:
          type: string
          description: 錯誤訊息
          example: 發生錯誤
        code:
          type: string
          description: 錯誤代碼
          example: ERROR_CODE
        timestamp:
          type: string
          format: date-time
          description: 回應時間
          example: '2024-01-01T00:00:00.000Z'
  responses:
...
    InternalServerError:
      description: 伺服器內部錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: 伺服器發生未預期錯誤
            code: INTERNAL_SERVER_ERROR
            timestamp: '2024-01-01T00:00:00.000Z'
...
paths:
  /auth/login:
    post:
      tags:
        - FrontAuth
      summary: 前台登入
      description: |
        以 Email 與密碼換發**前台的** Access Token 與 Refresh Token。

        **前台與後台是兩套獨立的帳號體系**：不同的表（`users` vs `members`）、
        不同的簽發 secret、不同的上下文型別。這裡簽出的 token
        **無法通過任何 `/api/admin/*` 端點的認證**——不是因為權限不足，
        而是簽章根本驗不過。

        登入失敗不區分「帳號不存在」與「密碼錯誤」，且帳號不存在時仍執行一次 bcrypt——
        少了那一步，兩條路徑的耗時差距穩定可測，足以用來列舉帳號。

        **不實作帳號鎖定。** 暴力破解的防護由全域 throttle 與 IP 封鎖負責——
        per-IP 而非 per-account，後者是一個未認證者可以觸發的 DoS 面。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
            example:
              email: user1@test.com
              password: User1234!
      responses:
        '200':
          description: 登入成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - timestamp
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required:
                      - accessToken
                      - refreshToken
                      - accessTokenExpiresIn
                      - refreshTokenExpiresIn
                      - user
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      accessTokenExpiresIn:
                        type: integer
                        description: 秒
                      refreshTokenExpiresIn:
                        type: integer
                        description: 秒
                      user:
                        type: object
                        required:
                          - id
                          - email
                          - displayName
                          - avatarUrl
                          - emailVerified
                        properties:
                          id:
                            type: string
                            format: uuid
                          email:
                            type: string
                            format: email
                          displayName:
                            type: string
                          avatarUrl:
                            type: string
                            nullable: true
                          emailVerified:
                            type: boolean
                            description: |
                              信箱是否已驗證。**false 時登得進來但聊不了天**——
                              客戶端要據此決定是引導去聊天室還是引導去收驗證信。
                              少了它，前台只能等第一次聊天請求被 403 才知道
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: email 格式不合法或 password 缺漏
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 帳號不存在或密碼錯誤（兩者訊息一致）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 帳號或密碼錯誤
                code: UNAUTHORIZED
                timestamp: '2026-08-23T06:00:00.000Z'
        '403':
          description: 帳號已停權
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 帳號已停用
                code: ACCOUNT_DISABLED
                timestamp: '2026-08-23T06:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/register:
    post:
      tags:
        - FrontAuth
      summary: 前台註冊
      description: |
        建立一個**信箱尚未驗證**的前台使用者，並寄出驗證信。無需認證。

        **本端點刻意會揭露「這個信箱是否已註冊」（409）**，與底下的重發驗證信
        和忘記密碼不同——那兩支一律回 204。不這樣做的話使用者收不到任何有用的回饋，
        他會以為註冊成功然後永遠等不到信。而「這個信箱能不能註冊」本來就是
        註冊表單必須回答的問題，藏不住。要擋的是**把它自動化**，那是限流的工作。

        **已註冊但尚未驗證的信箱再次註冊**：回 409 並**重發一次驗證信到原信箱**，
        不覆蓋既有帳號的任何欄位（尤其是密碼）。這是最常見的真實情境——
        信進了垃圾信匣，於是使用者重新註冊一次。擋掉他等於逼他換信箱。

        **回應不含 token**：註冊不等於登入，使用者要自己走登入流程。

        信箱在比對與儲存前會正規化（去空白、轉小寫），
        因此 `Foo@X.com` 與 `foo@x.com` 是同一個帳號。

        寄信**不阻塞回應**：SMTP 連不上時走滿逾時會讓註冊看起來當掉。
        寄送失敗只寫 log，不讓註冊本身失敗。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - displayName
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
                password:
                  type: string
                  maxLength: 128
                  description: 複雜度由密碼政策判定
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 50
            example:
              email: user@example.com
              password: User1234!
              displayName: 小明
      responses:
        '201':
          description: 註冊成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - timestamp
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required:
                      - id
                      - email
                      - displayName
                      - avatarUrl
                      - emailVerified
                    properties:
                      id:
                        type: string
                        format: uuid
                      email:
                        type: string
                        format: email
                      displayName:
                        type: string
                      avatarUrl:
                        type: string
                        nullable: true
                      emailVerified:
                        type: boolean
                        description: 註冊當下必然為 false
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: 欄位不符 schema，或密碼未通過政策
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 資料驗證失敗
                code: VALIDATION_ERROR
                timestamp: '2026-08-24T06:00:00.000Z'
        '409':
          description: 該信箱已註冊（未驗證者會順便重發驗證信）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 此 Email 已被使用
                code: EMAIL_ALREADY_EXISTS
                timestamp: '2026-08-24T06:00:00.000Z'
        '429':
          description: 對該信箱的寄送過於頻繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 寄送過於頻繁，請稍後再試
                code: EMAIL_SEND_RATE_LIMITED
                timestamp: '2026-08-24T06:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/verify-email:
    get:
      tags:
        - FrontAuth
      summary: 信箱驗證
      description: |
        驗證信裡的連結指向這裡。驗完 **302 導回前台**：
        `{APP_FRONT_URL}{APP_FRONT_VERIFY_REDIRECT_PATH}?result=success|invalid|expired`。

        **GET 且帶副作用，這是刻意的**——信件裡只能放連結，而連結只有 GET。
        代價是**預抓（prefetch）與郵件安全掃描會提前把 token 用掉**，
        因此**驗證成功是冪等的**：已驗證的帳號再次帶同一個 token 進來仍回
        `result=success`，而不是失敗。使用者什麼都沒做錯，不該看到失敗。

        **本端點不回 JSON 錯誤**：使用者是從信件點進來的，看到一段 JSON
        只會不知道發生什麼事。失敗一樣是 302，只是 `result` 不同。

        無效、過期、用途不符（拿密碼重設的 token 來驗證信箱）一律導向失敗，
        且不透露帳號是否存在。
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            maxLength: 255
          description: 信件連結帶入的明文 token
      responses:
        '302':
          description: |
            一律導回前台，以 query 的 `result` 表示結果：

            - `success` — 驗證完成（或本來就已驗證）
            - `expired` — token 過期
            - `invalid` — token 不存在、用途不符，或帳號已不存在
          headers:
            Location:
              schema:
                type: string
              description: '例：https://front.example.com/verify-email?result=success'
        '400':
          description: 未帶 token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 資料驗證失敗
                code: VALIDATION_ERROR
                timestamp: '2026-08-24T06:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/resend-verification:
    post:
      tags:
        - FrontAuth
      summary: 重發驗證信
      description: |
        對尚未驗證的帳號重發驗證信。無需認證。

        **無論信箱是否存在、是否已驗證，一律回 204。** 這一支與註冊不同：
        註冊揭露信箱狀態是為了給使用者有用的回饋，而重發沒有那個需求——
        它若依帳號狀態回不同的東西，就是一個乾淨的帳號探測點。

        已驗證的帳號不會收到信（沒有意義），但回應仍是 204。

        新的 token 會**作廢先前未使用的驗證 token**：使用者手上永遠只有最後一封有效，
        否則重發三次之後有三個同時有效的 token，而他只會用最新那個。

        限流的額度**無論帳號存不存在都會扣**——只對存在的帳號計數的話，
        Redis 的計數狀態就變成「這個信箱有沒有註冊」的旁通道。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
            example:
              email: user@example.com
      responses:
        '204':
          description: 已受理（無回應內容）。帳號不存在或已驗證時同樣回 204
        '400':
          description: 欄位不符 schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 資料驗證失敗
                code: VALIDATION_ERROR
                timestamp: '2026-08-24T06:00:00.000Z'
        '429':
          description: 對該信箱的寄送過於頻繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 寄送過於頻繁，請稍後再試
                code: EMAIL_SEND_RATE_LIMITED
                timestamp: '2026-08-24T06:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/forgot-password:
    post:
      tags:
        - FrontAuth
      summary: 前台忘記密碼
      description: |
        產生密碼重設 token 並寄出重設信。無需認證。

        **無論信箱是否存在一律回 204**，判準與重發驗證信相同——
        這一支沒有「給使用者有用的回饋」這個需求可以拿來抵消帳號列舉的風險：
        使用者本來就只會看到「信寄出去了，請去收信」。

        **未驗證信箱的帳號也可以用**：忘記密碼與信箱驗證是兩件事，
        而重設信本身就會送到那個信箱——能收到就證明他擁有它。
        擋掉未驗證者只會讓「註冊完忘記密碼」變成一個死結。

        新的 token 會作廢先前未使用的重設 token。

        限流的額度**無論帳號存不存在都會扣**，理由同重發驗證信。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
            example:
              email: user@example.com
      responses:
        '204':
          description: 已受理（無回應內容）。帳號不存在時同樣回 204
        '400':
          description: 欄位不符 schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 資料驗證失敗
                code: VALIDATION_ERROR
                timestamp: '2026-08-24T06:00:00.000Z'
        '429':
          description: 對該信箱的寄送過於頻繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 寄送過於頻繁，請稍後再試
                code: EMAIL_SEND_RATE_LIMITED
                timestamp: '2026-08-24T06:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/reset-password:
    post:
      tags:
        - FrontAuth
      summary: 前台重設密碼
      description: |
        以重設信裡的 token 設定新密碼。無需認證。

        成功會**遞增 `tokenVersion`，讓所有裝置立即登出**。那不是附加功能：
        會走到「忘記密碼」的情境本來就包含「帳號可能正被別人用著」，
        改完密碼卻讓對方的既有 session 繼續有效，等於只重設了一半。
        使用者需要用新密碼重新登入。

        **無效、過期、已使用、用途不符（例如拿驗證信的 token 來改密碼）
        一律回同一個錯誤**，四者不可區分——可區分的話就足以確認某個 token 曾經存在。

        密碼政策在消費 token 之前驗：反過來的話，密碼不合格時 token 已經沒了，
        使用者得回去重新申請一次。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
                - password
              properties:
                token:
                  type: string
                  maxLength: 255
                password:
                  type: string
                  maxLength: 128
                  description: 複雜度由密碼政策判定
            example:
              token: 3f6c1b2a8d4e4a9fb1c72e5d9a0f7b31
              password: NewPass1234!
      responses:
        '204':
          description: 已重設（無回應內容）。舊的 access token 同時失效
        '400':
          description: token 無效／過期／已使用／用途不符，或新密碼未通過政策
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: 連結無效或已過期，請重新申請
                code: INVALID_TOKEN
                timestamp: '2026-08-24T06:00:00.000Z'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags:
        - FrontAuth
      summary: 前台 Token 更新
      description: |
        以 Refresh Token 換發新的一對 Token（rotation）。
        舊的 Refresh Token 在換發成功後**立即進黑名單**——同一枚再次被使用
        是 token 遭竊的訊號，不可靜默放行。

        以 access token 呼叫、或以**後台**簽出的 refresh token 呼叫，都回 `401`
        （後者連簽章都驗不過，因為兩側用不同的 secret）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: 換發成功
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                  - timestamp
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/paths/~1auth~1login/post/responses/200/content/application~1json/schema/properties/data'
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: token 無效、過期、型別錯誤、側別錯誤或版本已撤銷
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 帳號在這期間被停權
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - FrontAuth
      summary: 前台登出
      description: |
        把當前的 access token 加入黑名單。

        **對客戶端是冪等的**：token 已經過期或無效時仍回 `204`——
        要求「先通過認證才能登出」會讓客戶端陷入「登不出去」的狀態。

        黑名單沿用後台的同一套實作：它以 token 本身為鍵，與哪一側簽發無關。
      responses:
        '204':
          description: 已登出
        '401':
          description: 未帶 Authorization header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...